4.0.0 Stability - Clean compilation warnings - Adjust PathNav level descriptor item UI to account for only one action button (as it never has children) - Add empty state when UI config entries are empty
