+++ ui/app/src/components/UsersGrid/styles.ts
@@ -16,7 +16,7 @@
 
 import { createStyles, makeStyles } from '@material-ui/core/styles';
 
-export const styles = makeStyles((theme) =>
+export const useStyles = makeStyles(() =>
   createStyles({
     root: {
       display: 'flex',
@@ -34,3 +34,5 @@ export const styles = makeStyles((theme) =>
     }
   })
 );
+
+export default useStyles;
