+++ ui/app/src/services/configuration.ts
@@ -24,6 +24,7 @@ import ContentInstance from '../models/ContentInstance';
 import { VersionsResponse } from '../models/Version';
 import LookupTable from '../models/LookupTable';
 import GlobalState from '../models/GlobalState';
+import { defineMessages } from 'react-intl';
 
 type CrafterCMSModules = 'studio' | 'engine';
 
@@ -121,6 +122,17 @@ export function setActiveTargetingModel(data): Observable<ActiveTargetingModel> 
 
 // endregion
 
+const messages = defineMessages({
+  emptyUiConfigMessageTitle: {
+    id: 'emptyUiConfigMessageTitle.title',
+    defaultMessage: 'Configuration is empty'
+  },
+  emptyUiConfigMessageSubtitle: {
+    id: 'emptyUiConfigMessageTitle.subtitle',
+    defaultMessage: 'Nothing is set to be shown here.'
+  }
+});
+
 export function fetchSiteUiConfig(site: string): Observable<Pick<GlobalState['uiConfig'], 'preview' | 'launcher'>> {
   return fetchConfigurationDOM(site, '/ui.xml', 'studio').pipe(
     map((xml) => {
@@ -128,10 +140,28 @@ export function fetchSiteUiConfig(site: string): Observable<Pick<GlobalState['ui
         const config = {
           preview: {
             toolsPanel: {
-              widgets: []
+              widgets: [
+                {
+                  id: 'craftercms.component.EmptyState',
+                  uiKey: -1,
+                  configuration: {
+                    title: messages.emptyUiConfigMessageTitle,
+                    subtitle: messages.emptyUiConfigMessageSubtitle
+                  }
+                }
+              ]
             },
             pageBuilderPanel: {
-              widgets: []
+              widgets: [
+                {
+                  id: 'craftercms.component.EmptyState',
+                  uiKey: -1,
+                  configuration: {
+                    title: messages.emptyUiConfigMessageTitle,
+                    subtitle: messages.emptyUiConfigMessageSubtitle
+                  }
+                }
+              ]
             }
           },
           launcher: null
